# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Randell Barber",
        "description": "Documentación de la API para la aplicación móvil Randell Barber. Incluye gestión de citas, pagos y administración.",
        "contact": {
            "email": "soporte@randellbarber.com"
        },
        "license": {
            "name": "Licencia Privada"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/appointments": {
            "get": {
                "summary": "Obtener lista consolidada de citas con filtros de fecha y estado. (Solo Admin/Barbero)",
                "operationId": "d57a3744531af775602d9beb75f44061",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Fecha de inicio (YYYY-MM-DD)"
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Fecha de fin (YYYY-MM-DD)"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado (confirmed, canceled, etc.)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de citas obtenida"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/dashboard-stats": {
            "get": {
                "summary": "Obtener estadísticas consolidadas del dashboard (Ingresos, conteo de estados). (Solo Admin/Barbero)",
                "operationId": "5c51275e9e102398c786a96484150345",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Fecha de inicio (YYYY-MM-DD)"
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Fecha de fin (YYYY-MM-DD)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estadísticas obtenidas"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}